backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Validation
pydantic==2.5.0

# Environment management
python-dotenv==1.0.0

# Development/Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2